fix warnings

+++ src/framework/animation/solvers/jacobian_solver.cpp
@@ -41,7 +41,7 @@ void JacobianSolver::set_rotation_axis()
 
 bool JacobianSolver::solve(const Transform& target)
 {
-    uint32_t num_joints = ik_chain.size() - 1u; // we don't need to compute the rotations for the end effector jont
+    uint32_t num_joints = static_cast<uint32_t>(ik_chain.size()) - 1u; // we don't need to compute the rotations for the end effector jont
     glm::vec3 goal_pos = target.get_position();
 
     // init the jacobi matrix (DoF * rotation axis)
